api/requirements.txt
fastapi
uvicorn[standard]
websockets
google-cloud-storage
python-multipart
